+++ cmd/builder/internal/builder/config.go
@@ -15,7 +15,7 @@ import (
 	"go.uber.org/zap"
 )
 
-const defaultOtelColVersion = "0.86.0"
+const defaultOtelColVersion = "0.87.0"
 
 // ErrInvalidGoMod indicates an invalid gomod
 var ErrInvalidGoMod = errors.New("invalid gomod specification for module")