+++ component/componenttest/nop_telemetry.go
@@ -21,5 +21,8 @@ func NewNopTelemetrySettings() component.TelemetrySettings {
 		MeterProvider:  noop.NewMeterProvider(),
 		MetricsLevel:   configtelemetry.LevelNone,
 		Resource:       pcommon.NewResource(),
+		ReportComponentStatus: func(*component.StatusEvent) error {
+			return nil
+		},
 	}
 }