+++ pdata/pmetric/generated_exemplarslice_test.go
@@ -11,13 +11,15 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"go.opentelemetry.io/collector/pdata/internal"
 	otlpmetrics "go.opentelemetry.io/collector/pdata/internal/data/protogen/metrics/v1"
 )
 
 func TestExemplarSlice(t *testing.T) {
 	es := NewExemplarSlice()
 	assert.Equal(t, 0, es.Len())
-	es = newExemplarSlice(&[]otlpmetrics.Exemplar{})
+	state := internal.StateMutable
+	es = newExemplarSlice(&[]otlpmetrics.Exemplar{}, &state)
 	assert.Equal(t, 0, es.Len())
 
 	emptyVal := NewExemplar()
@@ -29,6 +31,19 @@ func TestExemplarSlice(t *testing.T) {
 		assert.Equal(t, testVal, es.At(i))
 	}
 	assert.Equal(t, 7, es.Len())
+}
+
+func TestExemplarSliceReadOnly(t *testing.T) {
+	sharedState := internal.StateReadOnly
+	es := newExemplarSlice(&[]otlpmetrics.Exemplar{}, &sharedState)
+	assert.Equal(t, 0, es.Len())
+	assert.Panics(t, func() { es.AppendEmpty() })
+	assert.Panics(t, func() { es.EnsureCapacity(2) })
+	es2 := NewExemplarSlice()
+	es.CopyTo(es2)
+	assert.Panics(t, func() { es2.CopyTo(es) })
+	assert.Panics(t, func() { es.MoveAndAppendTo(es2) })
+	assert.Panics(t, func() { es2.MoveAndAppendTo(es) })
 }
 
 func TestExemplarSlice_CopyTo(t *testing.T) {
@@ -117,6 +132,6 @@ func fillTestExemplarSlice(es ExemplarSlice) {
 	*es.orig = make([]otlpmetrics.Exemplar, 7)
 	for i := 0; i < 7; i++ {
 		(*es.orig)[i] = otlpmetrics.Exemplar{}
-		fillTestExemplar(newExemplar(&(*es.orig)[i]))
+		fillTestExemplar(newExemplar(&(*es.orig)[i], es.state))
 	}
 }