+++ CHANGELOG.md
@@ -7,6 +7,18 @@ If you are looking for developer-facing changes, check out [CHANGELOG-API.md](./
 
 <!-- next version -->
 
+## v1.0.0-rcv0016/v0.87.0
+
+### 💡 Enhancements 💡
+
+- `service/telemetry exporter/exporterhelper`: Enable sampling logging by default and apply it to all components. (#8134)
+  The sampled logger configuration can be disabled easily by setting the `service::telemetry::logs::sampling::enabled` to `false`.
+- `core`: Adds the ability for components to report status and for extensions to subscribe to status events by implementing an optional StatusWatcher interface. (#7682)
+
+### 🧰 Bug fixes 🧰
+
+- `telemetry`: remove workaround to ignore errors when an instrument includes a `/` (#8346)
+
 ## v1.0.0-rcv0015/v0.86.0
 
 ### 🚩 Deprecations 🚩